+++ cmake/ImportIntegralEvaluator.cmake
@@ -3,35 +3,34 @@
 # Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
 # See LICENSE.txt for details.
 #
-macro(import_utils_os)
+macro(import_integral_evaluator)
   # If the target already exists, do nothing
-  if(NOT TARGET Scine::UtilsOS)
+  if(NOT TARGET Scine::IntegralEvaluator)
     # Try to find the package locally
-    find_package(ScineUtilsOS QUIET)
-    if(TARGET Scine::UtilsOS)
-      message(STATUS "Scine::UtilsOS found locally at ${ScineUtilsOS_DIR}")
+    find_package(ScineIntegralEvaluator QUIET)
+    if(TARGET Scine::LibintIntegrals)
+      message(STATUS "Scine::LibintIntegrals found locally at ${ScineIntegralEvaluator_DIR}")
     else()
       # Download it instead
       include(DownloadProject)
       download_project(
-        PROJ scine-utils-os
-        GIT_REPOSITORY https://github.com/qcscine/utilities.git
-        GIT_TAG        6.1.0
+        PROJ scine-integral-evaluator
+        GIT_REPOSITORY git@gitlab.chab.ethz.ch:scine/integralevaluator.git
+        GIT_TAG        1.0.0
         QUIET
       )
       # Note: Options defined in the project calling this function override default
       # option values specified in the imported project.
-      add_subdirectory(${scine-utils-os_SOURCE_DIR} ${scine-utils-os_BINARY_DIR})
-
+      add_subdirectory(${scine-integral-evaluator_SOURCE_DIR} ${scine-integral-evaluator_BINARY_DIR})
       # Final check if all went well
-      if(TARGET Scine::UtilsOS)
+      if(TARGET Scine::LibintIntegrals)
         message(STATUS
-          "Scine::UtilsOS was not found in your PATH, so it was downloaded."
+          "Scine::LibintIntegrals was not found in your PATH, so it was downloaded."
         )
       else()
         string(CONCAT error_msg
-          "Scine::UtilsOS was not found in your PATH and could not be established "
-          "through a download. Try specifying Scine_DIR or altering "
+          "Scine::LibintIntegrals was not found in your PATH and could not be "
+          "downloaded. Try specifying Scine_DIR or altering "
           "CMAKE_PREFIX_PATH to point to a candidate Scine installation base "
           "directory."
         )
@@ -40,3 +39,4 @@ macro(import_utils_os)
     endif()
   endif()
 endmacro()
+